requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==0.15.0
marshmallow==3.21.1
Flask-Limiter==3.6.0
Flask-Caching==2.1.0
pytest==8.2.0
requests==2.32.3
flask
python-jose[cryptography]
python-dotenv==1.0.1